move text formatting functions to separate module for testing without dependencies
