+++ server/kitsu/push.py
@@ -23,10 +23,12 @@ from .utils import (
     get_folder_by_kitsu_id,
     get_task_by_kitsu_id,
     get_user_by_kitsu_id,
-    remove_accents,
     update_folder,
     update_task,
 )
+
+from .addon_helpers import remove_accents
+
 
 if TYPE_CHECKING:
     from .. import KitsuAddon
