+++ server/kitsu/utils.py
@@ -37,30 +37,6 @@ def calculate_end_frame(
         return frame_start + nb_frames - 1
 
 
-def remove_accents(input_str: str) -> str:
-    nfkd_form = unicodedata.normalize("NFKD", input_str)
-    return "".join([c for c in nfkd_form if not unicodedata.combining(c)])
-
-
-def create_short_name(name: str) -> str:
-    code = name.lower()
-
-    if "_" in code:
-        subwords = code.split("_")
-        code = "".join([subword[0] for subword in subwords])[:4]
-    elif len(name) > 4:
-        vowels = ["a", "e", "i", "o", "u"]
-        filtered_word = "".join([char for char in code if char not in vowels])
-        code = filtered_word[:4]
-
-    # if there is a number at the end of the code, add it to the code
-    last_char = code[-1]
-    if last_char.isdigit():
-        code += last_char
-
-    return code
-
-
 def create_name_and_label(kitsu_name: str) -> dict[str, str]:
     """From a name coming from kitsu, create a name and label"""
     name_slug = slugify(kitsu_name, separator="_")
